% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.R
\name{estimate_beta}
\alias{estimate_beta}
\title{Estimate regression coefficients via M-estimation}
\usage{
estimate_beta(data_yXZ, model)
}
\arguments{
\item{data_yXZ}{A data frame containing the variables in \code{model}.
Must include the outcome \code{y} and covariates.}

\item{model}{An object of class \code{\link[stats]{formula}} specifying
the regression model for \code{y} in terms of covariates.}
}
\value{
A list with components
\describe{
\item{beta_est}{A 1 x (p + 1) matrix of estimated regression coefficients and
residual standard deviation (last element).}
\item{se_est}{A 1 x p matrix of estimated standard errors for the regression
coefficients.}
}
}
\description{
Estimate regression coefficients via M-estimation
}
