% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODRCC_PACKAGE.R
\docType{package}
\name{ODRCC}
\alias{ODRCC-package}
\alias{ODRCC}
\title{ODRCC: Robust Estimation Under \strong{O}utcome-\strong{D}ependent \strong{R}ight \strong{C}ensoring \strong{C}ovariate}
\description{
The ODRCC package implements robust and efficient estimation estimators for
regression models involving outcome-dependent right-censored covariates.
Available estimators inverse probability weighting (IPW),
augmented IPW (AIPW), and maximum likelihood (MLE).
}
\references{
Vazquez, J. et al. (2025). * Robust Estimation under Outcome Dependent Right Censoring
in Huntington Disease: Estimators for Low and High Censoring Rates*. Manuscript.
}
\author{
Jesus E. Vazquez
}
\keyword{internal}
